refactor: Member.Role 추가 및 entity 패키지 정리
