+++ src/main/java/roomescape/dto/reservation/ThemeRequest.java
@@ -1,6 +1,6 @@
 package roomescape.dto.reservation;
 
-import roomescape.entity.Theme;
+import roomescape.entity.reservation.Theme;
 
 public record ThemeRequest(
         String name,
