+++ src/main/java/roomescape/entity/member/Role.java
@@ -1,0 +1,5 @@
+package roomescape.entity.member;
+
+public enum Role {
+    USER, ADMIN
+}
