+++ src/main/java/roomescape/repository/member/MemberRepository.java
@@ -1,10 +1,12 @@
 package roomescape.repository.member;
 
-import roomescape.entity.Member;
+import roomescape.entity.member.Member;
 
 public interface MemberRepository {
 
     Member findById(Long id);
 
     boolean existsByEmailAndPassword(String email, String password);
+
+    Member findByEmail(String email);
 }
