+++ src/main/java/roomescape/service/memeber/UserReservationTimeService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.dto.reservation.ReservationAvailableTimeResponse;
-import roomescape.entity.ReservationTime;
+import roomescape.entity.reservation.ReservationTime;
 import roomescape.repository.reservation.ReservationRepository;
 import roomescape.repository.reservation.ReservationTimeRepository;
