+++ src/main/java/roomescape/service/reservation/ReservationService.java
@@ -5,10 +5,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.dto.reservation.ReservationRequest;
 import roomescape.dto.reservation.ReservationResponse;
-import roomescape.entity.Member;
-import roomescape.entity.Reservation;
-import roomescape.entity.ReservationTime;
-import roomescape.entity.Theme;
+import roomescape.entity.member.Member;
+import roomescape.entity.reservation.Reservation;
+import roomescape.entity.reservation.ReservationTime;
+import roomescape.entity.reservation.Theme;
 import roomescape.exceptions.reservation.ReservationDuplicateException;
 import roomescape.repository.member.MemberRepository;
 import roomescape.repository.reservation.ReservationRepository;
