+++ src/test/java/roomescape/entity/reservation/ThemeTest.java
@@ -1,4 +1,4 @@
-package roomescape.entity;
+package roomescape.entity.reservation;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
